dir: support text direction with .>>, .<<, \> and \<
The following troff requests specify text direction:

* .<< and .>> change the direction of text (.<< for r2l and .>>
  to l2r); the default direction is l2r.
* \< and \> change text direction temporarily.

Text direction processing starts after the first invocation
of .>> or .<<.

As an example, one can write:

  .<<
  This is a paragraph in a right-to-left language
  with some embedded \>left-to-right words\< in it.
  .>>

The current text direction and temporary direction are available
through .td and .cd number registers respectively; 0 for l2r and 1 for
r2l.

To indent the right side of output lines, two new requests have been
added: .in2 and .ti2 are equivalent to .in and .ti but for the right
side.  The .I number register stores the current right-side
indentation.