font: new syntax for defining characters and kerning pairs
This patch introduces a new syntax for defining characters or kerning
pairs: characters can be defined with lines starting with the word
"char" followed by the definition of a character similar to character
definitions in the charset section.  The lines starting with "kern"
followed by a triple (first glyph name, second glyph name, kerning
amount) describes a kerning pair.  Note that each kerning pair is
defined using glyph names (the fourth fields in character definitions)
and not character names as in the old, and now unsupported, kernpairs
section.  Note that the charset section is still supported.