cp: support conditional interpolation with \?'cond@expr1@expr2@'
This adds a new escape sequence for conditional interpolation: the
escape sequence \?'cond@expr1@expr2@', evaluates cond (exactly as if
it is a .if condition) and interpolates expr1, if the condition is
true, and expr2, otherwise.  The delimiter (@ in this example) can be
any character that cannot be part of the condition; for numerical
expressions, for instance, it cannot be a digit, an operator sign, or
a scale indicator, unless separated from the condition with \&.  The
final delimiter may be omitted.  Note that this escape sequence is not
interpolated in copy-mode.