font: support gsub/gpos context glyphs
For instance, the line

  gsub feat 5 =a -x -y -z +xyz

tells neatroff to replace the glyphs x y z with glyph xyz only if they
follow an a and font feature feat is enabled.  This patch also
supports matching one of the many glyphs in the patterns.  Also the
rule

  gsub feat 5 -a |b |c -y +axy

matches any of the strings ay, by and cy; neatroff replaces any
of them with the glyph axy if feat is enabled.