cp: remove c_ni from strings not pushed via in_push()
In order to prevent the interpretation of escaped backslashes,
cp.c replaces the first backslash of \\ with c_ni.  This is
necessary for ren.c escape sequences like \\D.  However, when
a string is processed directly and not read from ren.c, like
the arguments of built-in requests, c_ni characters should be
removed.  Reported by Carsten Kunze <carsten.kunze@arcor.de>.