fmt: consider the number of spaces while filling paragraphs
This updates the function for calculating the cost of breaking words
into lines to consider the number of spaces in each line.  The main
motivation is that a line with many spaces can be stretched much
better than a line with few large words.

This also updates the meaning of the argument of .hyp request: Now
useful values are between 0 and probably 1000; an argument of 100
assigns to each hyphenation the cost of stretching the spaces of the
line one hundred percent.  The default value is zero.