dev: finding font position automatically for .fp request
In Neatroff, if instead of the position of the font to be mounted,
the second argument of .fp is a dash, the position of the font is
decided automatically: if a font with the same name is already
mounted, the same position is reused.  Otherwise the font is mounted
on the first available font position.