ren: interpret ad request parameter as a set of flags
Now the argument of the ad request is assumed to be a set of flags.
Therefore, they can be combined.  For instance,
.ad br
means to adjust both margins, but right-adjust lines shorter than
the current line length (the last line of paragraphs, for instance).
Three new flags have also been added for adjusting lines in
no-fill mode: L, R, C.